#1636 Deleting a user no longer accidentally deletes a team if they're not an admin themselves

+++ app/User.php
@@ -286,23 +286,32 @@ class User extends Authenticatable
         parent::boot();
 
         // Delete user properly if it gets deleted
-        static::deleting(function ($item) {
-            /** @var $item User */
-            $item->dungeonroutes()->delete();
-            $item->reports()->delete();
+        static::deleting(function (User $user) {
+            $user->dungeonroutes()->delete();
+            $user->reports()->delete();
 
-            $item->patreonUserLink()->delete();
+            $user->patreonUserLink()->delete();
 
-            foreach ($item->teams as $team) {
+            foreach ($user->teams as $team) {
                 /** @var $team Team */
-                $newAdmin = $team->getNewAdminUponAdminAccountDeletion($item);
-                if ($newAdmin !== null) {
-                    // Appoint someone else admin
-                    $team->changeRole(User::find($newAdmin->id), 'admin');
-                    // Remove ourselves from the team
-                    $team->removeMember($item);
-                } else {
-                    $team->delete();
+                if (!$team->isUserAdmin($user)) {
+                    continue;
+                }
+
+                /** @var $team Team */
+                try {
+                    $newAdmin = $team->getNewAdminUponAdminAccountDeletion($user);
+                    if ($newAdmin !== null) {
+                        // Appoint someone else admin
+                        $team->changeRole(User::find($newAdmin->id), 'admin');
+                        // Remove ourselves from the team
+                        $team->removeMember($user);
+                    } else {
+                        // There's no new admin to be appointed - delete the team instead
+                        $team->delete();
+                    }
+                } catch (\Exception $exception) {
+                    logger()->error($exception->getMessage());
                 }
             }
             return true;
